services/ingestion-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
google-cloud-pubsub==2.27.1